TOML File Support (#10)
Adds support for documentation of both local and non-local functions
using a sniff-test.toml file included at the root directory of a cargo
project. Integrates with the newly simplified docstring parsing to
support identical syntax to in-code sniff test documentation.

This is primarily done by creating an additional toml module within the
annotations module, and passing around a TomlAnnotation object to
functions that at some point need to check functions for docstrings,
along with the already passed TyCtxt. If needed, we could later make a
struct that encapsulates all the information needed to check functions
for docstrings, but currently these are passed separately.

There are a few very brief tests included in tests/toml which
demonstrate how the toml file works.

---------

Co-authored-by: AlexanderPortland <alexanderportland1@gmail.com>