Handle checking traits (#15)
Adds support for checking traits by automatically visiting reachable
associated functions and ensuring that the obligations created by a
function that's an implementation of a trait matches the obligations
listed on the trait definition.

This also adds support for a custom check function for each property,
and they're currently used to ensure that all functions annotated as
having `# Safety` requirements also use Rust's `unsafe` keyword.